Refactoring of type conversion

+++ src/test/java/be/ugent/idlab/knows/functions/agent/functionInstantiator/InstantiatorTest.java
@@ -30,18 +30,18 @@ public class InstantiatorTest {
 
         Method sum = instantiator.getMethod("http://example.org/sum");
         Object result = sum.invoke(null, 5, 8);
-        assertTrue(result instanceof Integer);
-        Integer resultValue = (Integer) result;
+        assertTrue(result instanceof Long);
+        Long resultValue = (Long) result;
         assertEquals("5 + 8 should be 13", 13, resultValue.intValue());
     }
 
     @Test
     public void testLoadSumWithNonPrimitives() throws ClassNotFoundException, NoSuchMethodException, InvocationTargetException, IllegalAccessException {
         Class<?> clazz = Class.forName("be.ugent.idlab.knows.functions.internalfunctions.InternalTestFunctions");
-        Method method = clazz.getMethod("sum", int.class, int.class);
-        Object result = method.invoke(null, Integer.valueOf(1), Integer.valueOf(3));
-        assertEquals("1 + 3 should be 4", 4, result);
+        Method method = clazz.getMethod("sum", long.class, long.class);
+        Object result = method.invoke(null, Long.valueOf(1), Long.valueOf(3));
+        assertEquals("1 + 3 should be 4", 4L, result);
         Class<?>[] genericParameterTypes = method.getParameterTypes();
-        assertArrayEquals(new Class<?>[]{int.class, int.class}, genericParameterTypes);
+        assertArrayEquals(new Class<?>[]{long.class, long.class}, genericParameterTypes);
     }
 }
